+++ packages/dev/CHANGELOG.md
@@ -1,8 +1,8 @@
 @lifterlms/dev CHANGELOG
 ========================
 
-UNRELEASED
-----------
+v0.1.0 - 2022-08-11
+-------------------
 
 + Updated: [**BREAKING**] During prerelease builds the `readme` command will now exit with code `0` and output a warning message instead of previous behavior: exit code `1` with an error message. 
 + Added: The `readme` command now merges additional merge codes as derived from the `parseMainFileMetadata()` utility function.